all done

+++ data/migrations/20200123143017_mydatabase.js
@@ -33,7 +33,7 @@ exports.up = function(knex) {
           .notNullable()
           .references('id')
           .inTable('ingredients')
-        tbl.integer('ingredient_quantity')
+        tbl.float('ingredient_quantity')
           .notNullable()
       })
   };
